Add files via upload
There are just notes for anyone that already been using Linux a lot and is good from the command line interface. I skip the parts on deploying to large cloud companies because if you need that its best to use their resources to guide you. Please read the README on how this was based on VMs deployment as you may know containers are currently the next phase but I talked about how K8, VM, and Container are the next phase of IT infrastructure. If you have any questions please DM me but this was mostly written from the point of view of notes that I will need to use in the future.